Fix buffer list being incomplete at times
The editor count is wrong in certain circumstances... so just remove it. Instead of having a buggy mutable state which requires thorough debugging, it's better to compute the editor count on-the-fly. Checking that the list is empty is still O(1). Nobody will notice a difference in performance, unless thousands of editors are open.