Redesign LineNumbersWidth
LineNumbersWidth used to keep track of the margin width so that 'docView', 'leftMargin' and 'editorView' could be resized using relative amounts. However, that design implied coupling the internal state of Scintilla with the state of the UI widgets through a mutable state.

Now it only takes care of reporting and mutating Scintilla's state. The widgets can then be resized regardless of their previous state.

Calls to 'TView::changeBounds' have been replaced with calls to 'TView::setBounds' and 'TView::drawView' is no longer invoked on 'frame'. This avoids redundant draws because 'updateMarginWidth' is only invoked from 'EditorWindow::redraw', which already takes care of drawing these views.